Bit of a tidy up & minor refactor

+++ src/IISExpress.ts
@@ -3,101 +3,97 @@ import * as process from 'child_process';
 import * as path from 'path';
 import * as settings from './settings';
 
-//External libraries
+// External libraries
 import { v4 as uuidv4 } from 'uuid';
 import * as iconv from 'iconv-lite';
 
 export interface IExpressArguments {
-	path?: string;
-	port?: number;
-	clr?: settings.clrVersion;
-	protocol?: settings.protocolType;
+	path: string;
+	port: number;
+	clr: settings.clrVersion;
+	protocol: settings.protocolType;
 }
 
-export class IIS {
+export class IISExpress {
 	private _iisProcess!: process.ChildProcessWithoutNullStreams;
 	private _iisPath: string;
 	private _iisAppCmdPath: string;
-	private _args: IExpressArguments;
+	private _args!: IExpressArguments;
 	private _output!: vscode.OutputChannel | null;
 	private _statusbar!: vscode.StatusBarItem;
 	private _statusMessage!: string;
 
-	constructor(iisPath: string, appCmdPath: string, args: IExpressArguments){
+	constructor(iisPath: string, appCmdPath: string){
 		this._iisPath = iisPath;
 		this._iisAppCmdPath = appCmdPath;
-		this._args = args;
 	}
 
-	public startWebsite(options?: settings.Isettings) {
+	public startWebsite(options: settings.Isettings) {
 
-		//Verify process not already running, so if we have a PID (process ID) it's running
-		if(this._iisProcess !== undefined){
-			//Display error message that it's already running
-			vscode.window.showErrorMessage('IISExpress is already running');
+		// Verify process not already running, so if we have a PID (process ID) it's running
+		if(this._iisProcess !== undefined && this._iisProcess.killed === false){
+			// Display error message that it's already running
+			vscode.window.showErrorMessage('IIS Express is already running',{modal: true});
 
-			//Stop the method/function from running
+			// Stop the method/function from running
 			return;
 		}
-
 
 		if(!options){
-			vscode.window.showErrorMessage('No options found');
+			vscode.window.showErrorMessage('No options found in .vscode/iisexpress.json',{modal: true});
 			return;
 		}
 
-        //Get IIS Port Number from config file
-        this._args.port = options.port;
+		this._args = {
+			// Get IIS Port Number from config file
+			port: options.port,
 
-		//Folder to run as the arg
-		this._args.path = options.path ? options.path : vscode.workspace.rootPath;
+			// Folder to run as the arg
+			path: options.path ? options.path : <string>vscode.workspace.rootPath,
 
-		//CLR version, yes there are still people on 3.5 & default back to v4 if not set
-		this._args.clr = options.clr ? options.clr : settings.clrVersion.v40;
+			// CLR version, yes there are still people on 3.5 & default back to v4 if not set
+			clr: options.clr ? options.clr : settings.clrVersion.v40,
 
-		//If no protocol set fallback to http as opposed to https
-		this._args.protocol = options.protocol ? options.protocol : settings.protocolType.http;
+			// If no protocol set fallback to http as opposed to https
+			protocol: options.protocol ? options.protocol : settings.protocolType.http
+		};
 
-		//Create output channel & show it
+		// Create output channel & show it
 		this._output = this._output || vscode.window.createOutputChannel('IIS Express');
 		this._output.show(vscode.ViewColumn.Three);
 
-		//Site name is the name of the workspace folder & GUID/UUID
-		//Need to append a UUID as could have two folders/sites with same name
-		var siteName = path.basename(vscode.workspace.rootPath as string) + "-" + uuidv4();
+		// Site name is the name of the workspace folder & GUID/UUID
+		// Need to append a UUID as could have two folders/sites with same name
+		const siteName = path.basename(vscode.workspace.rootPath as string) + "-" + uuidv4();
 
-		//If user is using HTTPS & port not in range of auto-approved port numbers (44300-44399)
-		//Then display an error & stop process
+		// If user is using HTTPS & port not in range of auto-approved port numbers (44300-44399)
+		// Then display an error & stop process
 		if(this._args.protocol === settings.protocolType.https && (this._args.port >= 44300 && this._args.port <=44399) === false){
-			//Using HTTPS but not using a port within the range that supports SSL
-			vscode.window.showErrorMessage('When using HTTPS you need to use ports 44300 - 44399 in .vscode/iisexpress.json');
+			// Using HTTPS but not using a port within the range that supports SSL
+			vscode.window.showErrorMessage('When using HTTPS you need to use ports 44300 - 44399 in .vscode/iisexpress.json',{modal: true});
 
-			//Stop the method/function from running
+			// Stop the method/function from running
 			return;
 		}
 
-
-		//Add the site to the config (which will invoke/run from iisexpress cmd line)
-		//Not done as async - so we wait until this command completes
+		// Add the site to the config (which will invoke/run from iisexpress cmd line)
+		// Not done as async - so we wait until this command completes
 		try {
 			process.execFileSync(this._iisAppCmdPath, ['add', 'site', `-name:${siteName}`, `-bindings:${this._args.protocol}://localhost:${this._args.port}`, `-physicalPath:${this._args.path}`]);
 		} catch (error) {
 			console.log(error);
 		}
 
-		//Based on the CLR chosen use the correct built in AppPools shipping with IISExpress
-		var appPool = this._args.clr === settings.clrVersion.v40 ? "Clr4IntegratedAppPool" : "Clr2IntegratedAppPool";
+		// Based on the CLR chosen use the correct built in AppPools shipping with IISExpress
+		const appPool = this._args.clr === settings.clrVersion.v40 ? "Clr4IntegratedAppPool" : "Clr2IntegratedAppPool";
 
-		//Assign the apppool to the site
-		//appcmd set app /app.name:Site-Staging-201ec232-2906-4052-a431-727ec57b5b2e/ /applicationPool:Clr2IntegratedAppPool
+		// Assign the apppool to the site
+		// appcmd set app /app.name:Site-Staging-201ec232-2906-4052-a431-727ec57b5b2e/ /applicationPool:Clr2IntegratedAppPool
 		try {
 			process.execFileSync(this._iisAppCmdPath, ['set', 'app', `/app.name:${siteName}/`, `/applicationPool:${appPool}`]);
 		} catch (error) {
 			console.log(error);
 		}
-
-
-
 
 		// This is the magic that runs the IISExpress cmd from the appcmd config list
 		this._iisProcess = process.spawn(this._iisPath, [`-site:${siteName}`]);
@@ -131,7 +127,7 @@ export class IIS {
 		});
 
 		this._iisProcess.on('error', (err:Error) => {
-			var message = this.decode2gbk(err.message);
+			const message = this.decode2gbk(err.message);
 			this._output!.appendLine(`ERROR: ${message}`);
 		});
 
@@ -153,13 +149,16 @@ export class IIS {
 				console.log(error);
 			}
 		});
-
-
-		// Display Message
-		vscode.window.showInformationMessage(this._statusMessage);
 	}
 
 	public stopWebsite(){
+		// Ensure process is not null & is running
+		if(this._iisProcess === undefined || this._iisProcess.killed === true){
+			// Display error message
+			vscode.window.showErrorMessage('IIS Express is not running', {modal:true});
+			return;
+		}
+
 		// Kill the process - which will also hook into the exit event to remove the config entry
 		this._iisProcess.kill('SIGINT');
 
@@ -175,7 +174,7 @@ export class IIS {
 
 	}
 
-	public openWebsite(options?: settings.Isettings){
+	public openWebsite(options: settings.Isettings){
 
 		let browserConfig = vscode.workspace.getConfiguration().get<string>("iisexpress.openInBrowser", "default");
 		let browser:string = browserConfig.toLocaleLowerCase() === "default" ? "" : browserConfig;
@@ -192,22 +191,21 @@ export class IIS {
 		}
 	}
 
-	public restartSite(options? : settings.Isettings){
-		//If we do not have an iisProcess/website running
+	public restartSite(options : settings.Isettings){
+		// If we do not have an iisProcess/website running
 		if(!this._iisProcess){
-			//Then just do a start site...
+			// Then just do a start site...
 			this.startWebsite(options);
 		}
 		else {
-			//It's already running so stop it first then, start it
+			// It's already running so stop it first then, start it
 			this.stopWebsite();
 			this.startWebsite(options);
 		}
-
 	}
 
     private decode2gbk(data: string): string {
-		var buffer = new Buffer(data);
+		const buffer = new Buffer(data);
  		return iconv.decode(buffer, 'gbk');
 	}
